requirements.txt
psycopg2-binary==2.9.10
sqlalchemy==2.0.40
greenlet==3.1.1
